Hide `TransactionExecutor` from Python
The ultimate motivation is to make `BlockExecutor` a Facade, hiding all
Rust constructs from Python, except itself.
Next commits will hide `Storage`, completing the Facade.

All changes here are just making `PyBlockExecutor` swallow `PyTransactionExecutor`,
which no longer exists. This leaves the inner `TransactionExecutor` a
pure Rust struct.

Note: the transaction executor file name is still
`py_transaction_executor.rs`, so that we'll have a diff for this commit.
A subsequent commit will rename the file.
